Update some cython code to handle indices better
On Windows, the default Python 'integer' type is not a long,
but a long long. In general, this means we need to handle either
long or long long. For example, in the normals.pyx code, we use
fused_types to template the different integer types. We were
getting a failed test on windows due to not using fused_types
in the extract_patches method. However, fused_types don't play
well with the buffer interface.

Therefore, to make our code a bit more robust to very large
indices, I've changed from int/long to long long for indices.
This means taking more memory, but allows us to create really
massive images without error. If this is a real problem in terms
of memory then we can investigate proper templating with fused
types at a later date.

Note I have not update LBP and HOG with these changes as these
files need properly updating anyway.

+++ menpo/feature/cpp/central_difference.cpp
@@ -1,22 +1,24 @@
 #include <stdio.h>
 #include "central_difference.h"
-#include "central_difference.h"
 
-static inline int SUB2IND(const int j, const int i, const int k, const int row_size, const int col_size, const int n_channels) {
+static inline long long SUB2IND(const long long j, const long long i, const long long k,
+                                const long long row_size, const long long col_size,
+                                const long long n_channels) {
     //return ((i + col_size * j) * n_channels) + k;
     return (i + col_size * j) + (row_size * col_size * k);
 }
 
-void central_difference(const double* in, const int rows, const int cols,
-                        const int n_channels, double* out) {
-    const int n_output_channels = n_channels * 2;
-    int output_index = 0;
+void central_difference(const double* in, const long long rows,
+                        const long long cols, const long long n_channels,
+                        double* out) {
+    const long long n_output_channels = n_channels * 2;
+    long long output_index = 0;
 
     #pragma omp for
-    for (int k = 0; k < n_channels; k += 1) {
+    for (long long k = 0; k < n_channels; k += 1) {
         // row-derivative
-        for (int i = 0; i < rows; i += 1) {
-            for (int j = 0; j < cols; j += 1) {
+        for (long long i = 0; i < rows; i += 1) {
+            for (long long j = 0; j < cols; j += 1) {
                 if (j == 0) {
                     output_index = SUB2IND(0, i, k, rows, cols, n_output_channels);
                     out[output_index] = in[SUB2IND(1, i, k, rows, cols, n_channels)] - in[SUB2IND(0, i, k, rows, cols, n_channels)];
@@ -33,8 +35,8 @@ void central_difference(const double* in, const int rows, const int cols,
         }
 
         // column-derivative
-        for (int j = 0; j < rows; j += 1) {
-            for (int i = 0; i < cols; i += 1) {
+        for (long long j = 0; j < rows; j += 1) {
+            for (long long i = 0; i < cols; i += 1) {
                 if (i == 0) {
                     output_index = SUB2IND(j, 0, n_channels + k , rows, cols, n_output_channels);
                     out[output_index] = in[SUB2IND(j, 1, k, rows, cols, n_channels)] - in[SUB2IND(j, 0, k, rows, cols, n_channels)];
