+++ menpo/feature/cpp/central_difference.h
@@ -1,2 +1,3 @@
-void central_difference(const double* input, const int rows, const int cols,
-                        const int n_channels, double* output);
+void central_difference(const double* input, const long long rows,
+                        const long long cols, const long long n_channels,
+                        double* output);
